refactor(api): split platform and graph intelligence handlers
